# server/common/api.yml
openapi: 3.0.1
info:
  title: my-mh-collection
  description: My music home collection
  version: 1.0.0
servers:
- url: /api/v1
...
paths:
...
  /artists/{id}:
    get:
      tags:
        - Artists
      parameters:
        - name: id
          in: path
          description: The id of the artist to retrieve
          required: true
          schema:
            type: string
      responses:
        200:
          description: Return the artist with the specified id
          content: {}
        404:
          description: Artist not found
          content: {}
    delete:
      tags:
        - Artists
      parameters:
        - name: id
          in: path
          description: The id of the artist to delete
          required: true
          schema:
            type: string
      responses:
        204:
          description: Artist deleted
          content: {}
        404:
          description: Artist not found
          content: {}
...